pyproject.toml
[project]
name = "typesense_aio"
version = "0.1.0"
description = "Async client for Typesense"
authors = [
  { name = "devtud", email = "devtud@gmail.com" },
  { name = "trollfot", email = "trollfot@gmail.com" }
]
keywords = ["asyncio", "typesense"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Framework :: AsyncIO",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Software Development :: Libraries :: Application Frameworks"
]
requires-python = ">=3.9"
dependencies = [
    "httpx >= 0.25.0",
    "orjson",
]

[project.optional-dependencies]
test = [
     "docker >= 4.4.1",
     "black >= 20.8b1",
     "codecov >= 2.1.11",
     "pytest",
     "pyhamcrest",
     "alt-pytest-asyncio"
]

[tool.setuptools.packages.find]
where = ["."]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"